requirements.txt
Flask
Flask-Script
python-dotenv
Flask-Pymongo
pandas
bs4
requests
flask-mongoengine
flask-wtf
flask-login